+++ aegis_grpc/include/aegis_grpc/robot_control_service.hpp
@@ -66,6 +66,14 @@ class RobotControlServiceImpl final : public proto_aegis_grpc::v1::RobotControlS
                            const google::protobuf::Empty* request,
                            proto_aegis_grpc::v1::TriggerResponse* response) override;
 
+  grpc::Status WrenchBiasSet(grpc::ServerContext* context,
+                             const google::protobuf::Empty* request,
+                             proto_aegis_grpc::v1::TriggerResponse* response) override;
+
+  grpc::Status WrenchBiasClear(grpc::ServerContext* context,
+                               const google::protobuf::Empty* request,
+                               proto_aegis_grpc::v1::TriggerResponse* response) override;
+
  private:
   bool SwitchControllers(const std::vector<std::string>& activate, const std::vector<std::string>& deactivate);
   bool CallServoStartService();
@@ -91,6 +99,11 @@ class RobotControlServiceImpl final : public proto_aegis_grpc::v1::RobotControlS
   rclcpp::Publisher<geometry_msgs::msg::TwistStamped>::SharedPtr servo_tcp_pub_;
   rclcpp_action::Client<GripperCommand>::SharedPtr gripper_client_;
   rclcpp::TimerBase::SharedPtr servo_pub_timer_;
+
+  std::mutex wrench_bias_mutex_;
+  std::atomic<bool> wrench_bias_in_use_;
+  rclcpp::Client<std_srvs::srv::Trigger>::SharedPtr wrench_bias_set_client_;
+  rclcpp::Client<std_srvs::srv::Trigger>::SharedPtr wrench_bias_clear_client_;
 
   std::mutex servo_mutex_;
   ServoMode servo_mode_;